Moved Navigation Bar to separate file from App.js
